Configure travis to only build branch on master
Master builds in travis haven't been running for about 2 months. It
looks like it's because branch builds have been switched off in the UI.

To prevent branch builds on branches other that master the configuration
has been added. We don't need to build other branches because they are
captured in PR builds.