fixing log line for splunk (#135)
* Log exceptions in error handler

We should be catching and handling exceptions at the source where
applicable to resolve them in the best way. Where uncaught exceptions
slip through we weren't logging the traceback. This makes it difficult
to track the cause of the issue.

The error handler has been changed to log all tracebacks.