Replace generic 'RasError' and 'ClientError'
Nearly every exception raised in this service is a 'RasError'
or 'ClientError' exception (which is an extention of RasError).
I have removed these generic exceptions and replaced them with
the appropriate HTTPException.

Logging has also been improved as previously every RasError raised
was logged at error level and with the message 'Uncaught error'.